Update JE11853_first_Streamlit.py
Add my_cnx close() on Line#112 and Line#131
Since we open a Snowflake Connection each time a button on the page is clicked, let's also close the connection at the end of each button's script (line 58, below). We've illustrated the first one so that you can use it as an example and add a similar statement to the second button script. 

Finally, let's change a title and a button label. These just make a little more sense.